backend-auth/requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
lxml==5.3.2
python-docx==1.1.2
pywin32==310
tqdm==4.67.1
typing_extensions==4.13.2
